requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
pydantic[email]==2.5.0
python-multipart==0.0.6
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
tree-sitter==0.20.4
tree-sitter-languages==1.6.1